Rubrix/requirements.txt
# ===============================
# Core Backend
# ===============================
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
email-validator==2.1.1
python-dotenv==1.0.1

# ===============================
# Database (Relational)
# ===============================
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# ===============================
# Graph Database (Graph RAG)
# ===============================
neo4j==5.19.0
neomodel==5.3.0
networkx==3.3

# ===============================
# Auth & Security
# ===============================
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0

# ===============================
# Cache
# ===============================
redis==5.0.3

# ===============================
# LLM + RAG Framework (LOCKED VERSIONS)
# ===============================
langchain==0.1.16
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.0.8
langgraph==0.0.49

# ===============================
# Vector Database (Hybrid RAG)
# ===============================
pinecone-client==3.2.2
# qdrant-client==1.9.1
# faiss-cpu==1.8.0

# ===============================
# Embeddings & LLM Providers
# ===============================
openai==1.14.3
groq==0.4.2
sentence-transformers==2.6.1

# ===============================
# Observability & Evaluation
# ===============================
langsmith==0.1.17

# ===============================
# PDF Generation
# ===============================
Jinja2==3.1.3
WeasyPrint==61.0

# ===============================
# HTTP & Utils
# ===============================
httpx==0.27.0
requests==2.31.0
loguru==0.7.2
typing-extensions==4.11.0

# ===============================
# Testing
# ===============================
pytest==8.1.1
pytest-asyncio==0.23.6

#others
pymupdf==1.22.5
